# backend/schemata/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Test Observer",
    "description": "Test Observer API (see https://github.com/canonical/test_observer)",
    "license": {
      "name": "GNU Affero General Public License v3",
      "url": "https://raw.githubusercontent.com/canonical/test_observer/refs/heads/main/backend/LICENSE"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/issues": {
      "get": {
        "tags": [
          "issues"
        ],
        "summary": "Get Issues",
        "operationId": "get_issues_v1_issues_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by issue source (e.g., github, jira, launchpad)",
              "title": "Source"
            },
            "description": "Filter by issue source (e.g., github, jira, launchpad)"
          },
          {
            "name": "project",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by project name",
              "title": "Project"
            },
            "description": "Filter by project name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by issue status (e.g., open, closed, unknown)",
              "title": "Status"
            },
            "description": "Filter by issue status (e.g., open, closed, unknown)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of results to return (default: 50)",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of results to return (default: 50)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of results to skip for pagination (default: 0)",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of results to skip for pagination (default: 0)"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for issue source, project, keys, title, and status",
              "title": "Q"
            },
            "description": "Search term for issue source, project, keys, title, and status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuesGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "view_issue"
        ]
      },
      "put": {
        "tags": [
          "issues"
        ],
        "summary": "Create Or Update Issue",
        "operationId": "create_or_update_issue_v1_issues_put",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssuePutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_issue"
        ]
      }
    },
    "/v1/issues/{issue_id}": {
      "get": {
        "tags": [
          "issues"
        ],
        "summary": "Get Issue",
        "operationId": "get_issue_v1_issues__issue_id__get",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "view_issue"
        ]
      },
      "patch": {
        "tags": [
          "issues"
        ],
        "summary": "Patch Issue",
        "operationId": "patch_issue_v1_issues__issue_id__patch",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssuePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-permissions": [
          "change_issue"
        ]
      }
    },
...
    "/v1/auth/saml/login": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Saml Login",
        "operationId": "saml_login_v1_auth_saml_login_get",
        "parameters": [
          {
            "name": "return_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/saml/logout": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Saml Logout",
        "operationId": "saml_logout_v1_auth_saml_logout_get",
        "parameters": [
          {
            "name": "return_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecutionMetadata": {
        "additionalProperties": {
          "items": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "description": "Execution Metadata Value"
          },
          "type": "array"
        },
        "propertyNames": {
          "maxLength": 200,
          "minLength": 1,
          "description": "Execution Metadata Category"
        },
        "type": "object",
        "title": "Execution Metadata",
        "description": "A mapping of string categories to lists of string values.",
        "examples": [
          {
            "charm": [
              "postgresql-k8s",
              "mysql-k8s"
            ],
            "charm:postgresql-k8s:revision": [
              "495"
            ]
          }
        ]
      },
...
      "FamilyName": {
        "type": "string",
        "enum": [
          "snap",
          "deb",
          "charm",
          "image"
        ],
        "title": "FamilyName"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssuePatchRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "IssuePatchRequest"
      },
      "IssuePutRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "IssuePutRequest"
      },
      "IssueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/IssueSource"
          },
          "project": {
            "type": "string",
            "title": "Project"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          },
          "attachment_rules": {
            "items": {
              "$ref": "#/components/schemas/MinimalIssueTestResultAttachmentRuleResponse"
            },
            "type": "array",
            "title": "Attachment Rules"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "project",
          "key",
          "title",
          "status",
          "url",
          "attachment_rules"
        ],
        "title": "IssueResponse"
      },
      "IssueSource": {
        "type": "string",
        "enum": [
          "jira",
          "github",
          "launchpad"
        ],
        "title": "IssueSource"
      },
      "IssueStatus": {
        "type": "string",
        "enum": [
          "unknown",
          "open",
          "closed"
        ],
        "title": "IssueStatus"
      },
...
      "IssuesGetResponse": {
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/MinimalIssueResponse"
            },
            "type": "array",
            "title": "Issues"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "issues",
          "count",
          "limit",
          "offset"
        ],
        "title": "IssuesGetResponse"
      },
      "MinimalIssueResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/IssueSource"
          },
          "project": {
            "type": "string",
            "title": "Project"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "project",
          "key",
          "title",
          "status",
          "url"
        ],
        "title": "MinimalIssueResponse"
      },
...
      "MinimalIssueTestResultAttachmentRuleResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "families": {
            "items": {
              "$ref": "#/components/schemas/FamilyName"
            },
            "type": "array",
            "title": "Families"
          },
          "environment_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Environment Names"
          },
          "test_case_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Test Case Names"
          },
          "template_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Template Ids"
          },
          "test_result_statuses": {
            "items": {
              "$ref": "#/components/schemas/TestResultStatus"
            },
            "type": "array",
            "title": "Test Result Statuses"
          },
          "execution_metadata": {
            "$ref": "#/components/schemas/ExecutionMetadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "enabled",
          "families",
          "environment_names",
          "test_case_names",
          "template_ids",
          "test_result_statuses",
          "execution_metadata"
        ],
        "title": "MinimalIssueTestResultAttachmentRuleResponse"
      },
...
      "TestResultStatus": {
        "type": "string",
        "enum": [
          "PASSED",
          "FAILED",
          "SKIPPED"
        ],
        "title": "TestResultStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}